requirements.txt
python-dotenv==1.0.1
sounddevice==0.4.6
numpy==1.24.4
openwakeword==0.6.0
faster-whisper==0.10.0
TTS==0.22.0
requests==2.32.3
keyboard==0.13.5
fastapi==0.110.2
uvicorn==0.30.1
pytest==8.1.1
